backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
numpy==1.26.4
Pillow==10.4.0
python-multipart==0.0.9
